Use PBKDF2 key derivation for crypto password (#5)
* Use PBKDF2 key derivation for crypto password

The random IV is reused for the salt. This means the encryption key
won't be static anymore and changes for every message. It also makes
brute-forcing harder.

* Use separate keys for encryption and authentication

This helps against potential attacks that benefit from any similarities
between encryption and authentication algorithms.

* Remove password from MAC generation

Not very useful and might be potentially dangerous.